requirements.txt
cairocffi==1.6.1
CairoSVG==2.7.1
cffi==1.15.1
cssselect2==0.7.0
defusedxml==0.7.1
Pillow==9.0.1
pycparser==2.21
tinycss2==1.2.1
webencodings==0.5.1
